moved selector logic
