+++ src/redux/workout/workoutSlice.test.js
@@ -1,33 +1,6 @@
-import {
-  getSetCountFromWorkoutExerciseIdList,
-  getSetCountFromWorkoutId,
-} from './workoutExerciseSetSlice';
+import {getSetCountFromWorkoutId} from './workoutSlice';
 
 describe('selectors', () => {
-  describe('getSetCountFromWorkoutExerciseIdList', () => {
-    it('returns 0 with no items', () => {
-      const state = {
-        workoutExerciseSet: {
-          workoutExerciseSets: [],
-        },
-      };
-
-      const result = getSetCountFromWorkoutExerciseIdList(state, [1]);
-      expect(result).toEqual(0);
-    });
-
-    it('returns 1 with 1 item', () => {
-      const state = {
-        workoutExerciseSet: {
-          workoutExerciseSets: [{workout_exercise: 1}],
-        },
-      };
-
-      const result = getSetCountFromWorkoutExerciseIdList(state, [1]);
-      expect(result).toEqual(1);
-    });
-  });
-
   describe('getSetCountFromWorkoutId', () => {
     it('returns 0 with no items', () => {
       const state = {
